Move Invocation.count_invocation_reasons_for_user/1 to Accounts.has_activity_in_projects?/1
